+++ src/PowerShellEditorServices.Protocol/DebugAdapter/AttachRequest.cs
@@ -20,6 +20,10 @@ namespace Microsoft.PowerShell.EditorServices.Protocol.DebugAdapter
 
         public string ProcessId { get; set; }
 
-        public int RunspaceId { get; set; }
+        public string RunspaceId { get; set; }
+
+        public string RunspaceName { get; set; }
+
+        public string CustomPipeName { get; set; }
     }
 }