+++ src/PowerShellEditorServices/Analysis/AnalysisService.cs
@@ -52,11 +52,6 @@ namespace Microsoft.PowerShell.EditorServices
         /// </summary>
         private static readonly PSObject[] s_emptyDiagnosticResult = new PSObject[0];
 
-        /// <summary>
-        /// An empty script marker result to return when no script markers can be returned.
-        /// </summary>
-        private static readonly ScriptFileMarker[] s_emptyScriptMarkerResult = new ScriptFileMarker[0];
-
         private static readonly string[] s_emptyGetRuleResult = new string[0];
 
         /// <summary>
@@ -266,7 +261,7 @@ namespace Microsoft.PowerShell.EditorServices
         /// </summary>
         /// <param name="file">The ScriptFile which will be analyzed for semantic markers.</param>
         /// <returns>An array of ScriptFileMarkers containing semantic analysis results.</returns>
-        public async Task<ScriptFileMarker[]> GetSemanticMarkersAsync(ScriptFile file)
+        public async Task<List<ScriptFileMarker>> GetSemanticMarkersAsync(ScriptFile file)
         {
             return await GetSemanticMarkersAsync<string>(file, ActiveRules, SettingsPath);
         }
@@ -277,7 +272,7 @@ namespace Microsoft.PowerShell.EditorServices
         /// <param name="file">The ScriptFile to be analyzed.</param>
         /// <param name="settings">ScriptAnalyzer settings</param>
         /// <returns></returns>
-        public async Task<ScriptFileMarker[]> GetSemanticMarkersAsync(ScriptFile file, Hashtable settings)
+        public async Task<List<ScriptFileMarker>> GetSemanticMarkersAsync(ScriptFile file, Hashtable settings)
         {
             return await GetSemanticMarkersAsync<Hashtable>(file, null, settings);
         }
@@ -288,7 +283,7 @@ namespace Microsoft.PowerShell.EditorServices
         /// <param name="scriptContent">The script content to be analyzed.</param>
         /// <param name="settings">ScriptAnalyzer settings</param>
         /// <returns></returns>
-        public async Task<ScriptFileMarker[]> GetSemanticMarkersAsync(
+        public async Task<List<ScriptFileMarker>> GetSemanticMarkersAsync(
            string scriptContent,
            Hashtable settings)
         {
@@ -379,7 +374,7 @@ namespace Microsoft.PowerShell.EditorServices
 
         #region Private Methods
 
-        private async Task<ScriptFileMarker[]> GetSemanticMarkersAsync<TSettings>(
+        private async Task<List<ScriptFileMarker>> GetSemanticMarkersAsync<TSettings>(
             ScriptFile file,
             string[] rules,
             TSettings settings) where TSettings : class
@@ -394,11 +389,11 @@ namespace Microsoft.PowerShell.EditorServices
             else
             {
                 // Return an empty marker list
-                return s_emptyScriptMarkerResult;
+                return new List<ScriptFileMarker>();
             }
         }
 
-        private async Task<ScriptFileMarker[]> GetSemanticMarkersAsync<TSettings>(
+        private async Task<List<ScriptFileMarker>> GetSemanticMarkersAsync<TSettings>(
             string scriptContent,
             string[] rules,
             TSettings settings) where TSettings : class
@@ -407,12 +402,12 @@ namespace Microsoft.PowerShell.EditorServices
                 && (rules != null || settings != null))
             {
                 var scriptFileMarkers = await GetDiagnosticRecordsAsync(scriptContent, rules, settings);
-                return scriptFileMarkers.Select(ScriptFileMarker.FromDiagnosticRecord).ToArray();
+                return scriptFileMarkers.Select(ScriptFileMarker.FromDiagnosticRecord).ToList();
             }
             else
             {
                 // Return an empty marker list
-                return s_emptyScriptMarkerResult;
+                return new List<ScriptFileMarker>();
             }
         }
 
@@ -514,7 +509,9 @@ namespace Microsoft.PowerShell.EditorServices
                     new Dictionary<string, object>
                     {
                         { "ScriptDefinition", scriptContent },
-                        { settingParameter, settingArgument }
+                        { settingParameter, settingArgument },
+                        // We ignore ParseErrors from PSSA because we already send them when we parse the file.
+                        { "Severity", new [] { ScriptFileMarkerLevel.Error, ScriptFileMarkerLevel.Information, ScriptFileMarkerLevel.Warning }}
                     });
 
                 diagnosticRecords = result?.Output;